restaurant server

+++ src/main/java/com/rama/restaurant/repository/RestaurantRepository.java
@@ -1,12 +1,9 @@
 package com.rama.restaurant.repository;
-
 import com.rama.restaurant.entity.Restaurant;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
-
 @Repository
-public interface RestaurantRepository extends JpaRepository<Restaurant, Integer> {
-
+public interface RestaurantRepository extends JpaRepository<Restaurant, Long> {
 }
+
